+++ src/main/java/com/example/firstcrud/services/IEquipeService.java
@@ -12,5 +12,5 @@ public interface IEquipeService {
 
     Equipe updateEquipe (Equipe e);
 
-    Equipe retrieveEquipe (Integer idEquipe);
+    Equipe retrieveEquipe (Long idEquipe);
 }
